add checks for permissions to change page URL
Check whether the current Page allows changing URL, and whether the currently logged in user has the required permissions to change it. If either of these permissions is not present, then the page URL in the SEO field can now be safely clicked without doing anything, and without generating an error.